Create tableizer.js
This plugin will allow you to take a set of objects (array of objects, such as taken from APIs) and convert them into a formatted table, capable of sorting into order both forward and reverse by clicking on the "login" label. The search box allows one to filter list members. It also utilizes debounce, which will delay the narrowing of the list to account for typing speed.